refactor: comments modules refactor
