+++ src/comments/comments.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { CommentsController } from './controller/comments/comments.controller';
-import { CommentsService } from './services/comments/comments.service';
+import { CommentsController } from './controller/comments.controller';
+import { CommentsService } from './services/comments.service';
 import { SharedServicesModule } from 'src/shared/shared-services.module';
 
 @Module({
